Crate utils_lib_derive

Procedural macro: utils-lib-derive

Derive macro for Getter and Sealed trait.

This crate won’t be publish on crates.io. And I might introduce breaking change at anytime. For now I am not completely happy with the syntax of the Getter derive macro (but may require too much time for me). I have also plan to add more derive macro in the future.

Example

Getter

use utils_lib_derive::Getter;

#[derive(Getter)]
struct S {
    #[get(Pub, Const)]
    f: u32,
    #[get_mut(Pub)]
    f2: i8
}

fn main() {
    let mut s = S { f: 1_u32, f2: -1_i8 };
    assert_eq!(s.f(), &1_u32);
    assert_eq!(s.f2_mut(), &-1_i8);
}

Sealed

use utils_lib_derive::{trait_sealed, Sealed};

// this create a module named [`private`] with a trait named [`Sealed`]
// without method inside that module.
trait_sealed!();

#[derive(Sealed)]
struct S;

/// this trait is sealed and cannot me implemented outside of this crate
/// because [`Sealed`] is a private trait that can't be implemented outside
/// of this crate.
pub trait Trait: private::Sealed {}

impl Trait for S {}

fn main() {}

See more example

Macros

trait_sealed

Creates a trait Sealed into a private module private.

Derive Macros

Getter

Derive getter macro

Sealed

Derive the Sealed trait